Include variables and routes for authentication: - Move the 'isUserAuthenticated' function to authentication.js in auth folder - Add variables for authentication controllers and middleware - Modify post route to reference the new action in AuthController
